refactor: substrate and ethereum indexer endpoint.
