+++ src/endpoints/substrate-indexed-data/labs/lab.controller.ts
