+++ src/endpoints/substrate-indexed-data/labs/lab.service.ts
