+++ test/unit/substrate-indexed-data/labs/lab.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { when } from 'jest-when';
 import { ElasticsearchService } from '@nestjs/elasticsearch';
-import { LabService } from '../../../../src/substrate-indexed-data/labs/lab.service';
+import { LabService } from '../../../../src/endpoints/substrate-indexed-data/labs/lab.service';
 import { elasticsearchServiceMockFactory, MockType } from '../../mock';
 
 describe('Substrate Indexer Lab Service Unit Tests', () => {
